Add files via upload
The purpose of this file is to capture the accelerometer data in 14.28ms. The freuency used is 70Hz. The total sample capture in one second is 210. These are values of x y and z. The data is in time series. 
The approach used is real time hardware interrupt. The loop is going to run 210 times and it will calaculate the values. The approach used the hardware interrupt which is "Counter".